+++ frontend/js/gens.ts
@@ -65,7 +65,6 @@ export async function initCanvases({
   scoutBaseURL,
   gensApiURL,
   mainSampleTypes,
-  annotationFile: defaultAnnotationName,
   startRegion,
   version,
   allSamples,
@@ -174,21 +173,13 @@ export async function initCanvases({
   setupShortcuts(session, sideMenu, inputControls, onChromClick);
 
   const allAnnotSources = await api.getAnnotationSources();
-
-  const defaultAnnot = allAnnotSources
-    .filter((track) => track.name === defaultAnnotationName)
-    .map((track) => {
-      return {
-        id: track.track_id,
-        label: track.name,
-      };
-    });
+  const geneLists = await api.getGeneLists();
 
   settingsPage.setSources(
     session,
     render,
     allAnnotSources,
-    defaultAnnot,
+    geneLists,
     () => gensTracks.trackView.getDataTracks(),
     (trackId: string, direction: "up" | "down") =>
       gensTracks.trackView.moveTrack(trackId, direction),