+++ gens/routes/gene_lists.py
@@ -24,19 +24,5 @@ async def get_gene_list_symbols(
 ) -> list[str]:
     """Get gene list entries"""
 
-    gene_names = variant_adapter.get_panel(panel_id)
+    gene_names = variant_adapter.get_gene_list(panel_id)
     return gene_names
-    # if not gene_names:
-    #     return []
-
-    # all_matches: list[SimplifiedTranscriptInfo] = []
-    # for gene_name in gene_names:
-
-    #     matches = get_simplified_transcripts_by_gene_symbol(
-    #         gene_name, genome_build, db, only_mane=True
-    #     )
-
-    #     matches_in_chr = [tr for tr in matches if tr.chrom == chromosome.value]
-    #     all_matches.extend(matches_in_chr)
-
-    # return all_matches