+++ CHANGELOG.md
@@ -11,11 +11,15 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
  - ClinVar IDs (VCV, RCV) are detect in annotation text and automatically made into URLs. (https://github.com/SMD-Bioinformatics-Lund/gens/pull/512)
  - Add `--baf` and `--coverage` flags to `gens update sample` CLI command. (https://github.com/SMD-Bioinformatics-Lund/gens/pull/518)
  - Allow using sample types T/N as aliases for tumor/normal. They are stored internally as "tumor" and "normal". (https://github.com/SMD-Bioinformatics-Lund/gens/pull/518)
+ - IndexedDB caching of loaded transcripts (https://github.com/SMD-Bioinformatics-Lund/gens/pull/526)
+ - Allow displaying Scout panels as tracks in Gens (https://github.com/SMD-Bioinformatics-Lund/gens/pull/526)
 
 ### Changed
  - Add a small padding at sides when navigating to a gene or annotation through search. (https://github.com/SMD-Bioinformatics-Lund/gens/pull/516)
  - Default hide advanced settings behind a toggle button. (https://github.com/SMD-Bioinformatics-Lund/gens/pull/517)
  - Raise an error if attempting to load samples where the cov or baf files (and tbi files) are not present. (https://github.com/SMD-Bioinformatics-Lund/gens/pull/518)
+ - Removed default annotation track setting (https://github.com/SMD-Bioinformatics-Lund/gens/pull/526)
+ - Generalize connection with Scout using an abstract class as interface, opening for interfacing with other alternative variant software (https://github.com/SMD-Bioinformatics-Lund/gens/pull/526)
 
 ### Fixed 
  - Coloring of heterozygote variants now uses correct colors. (https://github.com/SMD-Bioinformatics-Lund/gens/pull/514)